functions/ingest/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery==3.14.0
google-cloud-storage==2.14.0
pandas==2.1.4
openpyxl==3.1.2
google-generativeai==0.3.2
thefuzz==0.20.0
PyYAML==6.0.1
